Fix GD&T CI import path
