+++ docs/kcl-std/types/std-types-EdgeEndpoint.md
@@ -6,3 +6,8 @@ layout: manual
 ---
 
 The start or end endpoint of an edge.
+
+
+
+
+
